smart-ml-preprocessor/requirements.txt
flask
kafka-python
pandas